+++ examples/ping_pong/main.rs
@@ -2,8 +2,10 @@ extern crate busan;
 
 use busan::actor::{Actor, ActorAddress, ActorInit, Context};
 use busan::config::ActorSystemConfig;
-use busan::message::ToMessage;
+use busan::message::common_types::{I32Wrapper, StringWrapper};
+use busan::message::{Message, ToMessage};
 use busan::system::ActorSystem;
+use std::any::Any;
 use std::thread;
 
 struct Ping {
@@ -14,7 +16,7 @@ struct Pong {
 }
 
 impl ActorInit for Ping {
-    type Init = ();
+    type Init = I32Wrapper;
 
     fn init(_init_msg: &Self::Init) -> Self
     where
@@ -26,7 +28,7 @@ impl ActorInit for Ping {
 }
 
 impl ActorInit for Pong {
-    type Init = ();
+    type Init = I32Wrapper;
 
     fn init(_init_msg: &Self::Init) -> Self
     where
@@ -39,11 +41,12 @@ impl ActorInit for Pong {
 
 impl Actor for Ping {
     fn before_start(&mut self, mut ctx: Context) {
-        self.pong_addr = Some(ctx.spawn_child::<_, Pong>("pong".to_string(), &()));
+        self.pong_addr =
+            Some(ctx.spawn_child::<_, Pong>("pong".to_string(), &I32Wrapper::default()));
         ctx.send_message(&self.pong_addr.as_ref().unwrap(), "ping".to_message());
     }
 
-    fn receive(&mut self, ctx: Context, _msg: Box<dyn prost::Message>) {
+    fn receive(&mut self, ctx: Context, _msg: Box<dyn Message>) {
         println!("received message");
         // assume it was a pong, send a ping
         match &self.pong_addr {
@@ -53,9 +56,13 @@ impl Actor for Ping {
     }
 }
 impl Actor for Pong {
-    fn receive(&mut self, ctx: Context, _msg: Box<dyn prost::Message>) {
+    fn receive(&mut self, ctx: Context, _msg: Box<dyn Message>) {
         println!("received message");
         // assume it was a ping, send a pong
+        match _msg.as_any().downcast_ref::<StringWrapper>() {
+            Some(msg) => println!("received message: {}", msg.value),
+            None => {}
+        }
         match &self.ping_addr {
             Some(addr) => ctx.send_message(&addr, "pong".to_message()),
             None => {}
@@ -65,7 +72,7 @@ impl Actor for Pong {
 
 fn main() {
     let mut system = ActorSystem::init(ActorSystemConfig::default());
-    system.spawn_root_actor::<_, Ping>("ping".to_string(), &());
+    system.spawn_root_actor::<_, Ping>("ping".to_string(), &I32Wrapper::default());
 
     thread::sleep(std::time::Duration::from_secs(1));
     system.shutdown();